UPDATE!!!!! D:<
So as you can see, a whole bunch has been messed with by me. The medic
doors are nto finished because I don't know where to go with them
color-wise. The map has gotten a big update. There is now a proper
Mechbay in place as well as a vault. Only two mechs on the map, ripley
and firefighter ripley, so make the others you want. Mech weapons are in
the weapons vault held by warden. Added landmark for Addict's clothes
andd placed in R&D office. Also changed the R&D PDA a little to match
his favorite color. Not a big change.
Lots of things don't work in here though so that will be a later fix.
Most likely the next one if we can find out what's wrong.