Getting up there...
Compared my edits to the lastest commit and pasted them over.
jobprocs.dm has updated code making a female clown or mime to spawn with
the "sexy" version of the type that exists (they aren't too sexy
looking). I redid the B.L.A.R.G. Fast Missile Rack and had to update it
in the Warthog also. Everything else is still a work in process for me.
I just have to find out what should be what, where, and how. Should be
done on next commit from me.