Paper Code...
I have managed to copy code from BayStation 12 that allows you to use
HTML/BBCode/whatever they use when writing something on paper. It WORKS.
I tested it with a black pen and purple crayon. Both work as they should
with the proper tags used. However commiting it isn't. This is a test
commit before I apply all the fixes for it. Don't use this if it works,