An update...
Okay so here's the run down. Made a Legend of Zelda shield in weapon.dm
and dmi as well in the left and right hands .dmi files. Going to find
out how to make my dual katanas in twohanded.dm. Fixed the clipboard
icon not showing. Am adding Warthog constructs so it can be built.
Deleted ALL the backup map files because they are crap and taking up
memory. They aren't needed anyways. Updated Blarg2 with a better Medbay.
Not sure if that's done completely but I did a good job in it. Better
than what was in it. Also adding in a PDA Fabricator so PDAs can be
made. Also I don't know why I added that in there but it seems like a
nice idea...

NEXT COMMIT WILL BE TOTAL DO OVER OF THE CODE. DON'T FREAK OUT.