ORGANIZED BRANCH!
Yeah, I have no idea if this is better than what is currently SS13 code,
but I'll be damned if it is worse. It's also not done yet. I moved a
large majority of all the files and procs/vars/stuff into files that
matched the objects tree. Some of them need to be more rearranged but
that can wait for a while. Right now here is something to show all my
work.