Part 2 of changes
This is the other half of the changes I intended to make today.
Changelog:

*Moved Shaft Miner spawn. Evidently, they don't have cargo access. Why,
is BYOND me. (may have fixed this issue below)
*Fixed a lightbulb bug in ruskie station. *slaps self*
!!!*FUCK IT. Ruskie station is now simulated turf.
-Deleted a slight amount of WEEEEED
+Added a switch to operate the disposal conveyor belt. Why did it get
removed?
+Put a water tank in janitor's closet... even though there was one in
the maintenance tunnel not but two tiles away. oh well.
+Added floors beneath the prison shuttle's doors; they had space beneath
them for whatever reason.
*Hopefully fixed the door access problems in QM.
+Added a toxins test chamber
!>Testing with area instances. Probably won't have any adverse effects
NOW, but it may in the future. I'm making a note here for future
reference.
!>Derelict mining station has an experimental use of disposal. Making a
note here, possible success.
*+Finished the derelict mining station to the best of my knowledge.
!!!*fixed that fucking arrival door.
+added a light in HoP's domain that they have needed for a while
-Removed money crates from the supply ordering console. I did not take
out any code; rather, I block commented it out. If there is ever a need
to use this code in the future, it's still there.