Bllleeaaaarrreegh
Yet another fragment, to be appended to the previous set of shit. I
found myself zoning out frequently while working in this, so I figured
it was time to call it a day and roll the intents for this update over
to the next time I bugger the code.

Changelog:
*Changed all "derelict mining" to "geostation". this is to avoid
confusion later on down the road.
*fixed paper not being renamable, despite the command being present.
*put a bit more of the station on the new disposal. I'll get the rest of
it eventually.
*fixed some air pipe fuckups in mining station
+added some lights in geostation to rooms that were missing them
!!!!!!!!!!!!!!!!!*FIXED. THAT. FUCKING. DOOR.
*Fixed a few other door access issues. Most common ones involve
defaulting the base external airlock model, which has a required access
of 13 (External access permission), which a lot of jobs didn't have. I
suggest from now on that you either not use the red airlocks for indoor
use or make sure to edit the access parameterS correctly.
+made the disposal system a little bit less hazardous by putting flaps
and windows in.

Notes:
FOR SOME FUCKING REASON, the new geostation area throws a fuckton of
errors in dream daemon, and I don't know why. I humbly request someone
take a look into this and tell me what I did wrong. Only side effect I
can see thus far is that the area in question is being powered when it
should not be. (The APC doesn't even have a cell in it and is hanging
wide open; this gimmick was working fine before I started fucking with
areas).