A critical unfail, plus other
So as I was loooking thru the code, I realized I forgot to include the
blarg2 map so we could play on it. This seems like it would be a bad
thing, so I fixed that.

I also did this:
*Disposal system fixed, for the most part. a few of the rooms on the
left side still eject into space. (This was actually done last commit,
but I forgot to log it)
-Removed stealing a functional AI from possible traitor objectives for
the time being.
+added another endgame soundboard