Fruits of my rage
Fun fact: I did this to distract myself so I didn't murder someone.
Literally.

*Pipes will autofling, evidently. Removed the ejection unit thing in the
waste room and replaced it with a different system. This may or may not
work.
+Missing pipes fixed in the QM's window.
*fixed a slew of other disposal pipe bugs I didn't even know existed
until now
*wallmounted several of the disposal units. Chill out tauka, not the way
you think I did.
+Added a disposal unit in surgery room. prior to this, there was only a
single disposal in the entire medbay.
*fixed some wiring fail
*fixed some air pipe fail
+Put an intercom in medbay. This is for future plans. Sneaky future
plans.
+QM and his/her subordinates have access to tech storage (sans miner,
and assuming they can get there) and emergency storage.
*fixed a door permission that caused mulebots to take a very roundabout
way to get to the beacon in engineering hall. This is a temporary
solution; I need to do something else to really 'fix' this.
*terralabs is now reminiscent of the terror it had long before.
-Removed a duplicate APC in geostation