urrrrrrrngh
did a bunch of random shit, some of which were things on my to-do list,
some of which weren't.

+added a bridge-specific instance of the SMES and replaced the ones
there already with the new ones
*reworded the ROCKBOAT verb
+added a personal space for clown and mime
+added areas to accomodate above
+added more noticeboards
*moved some lockers in library that were ignorantly placed
*Set the stacker to discharge stacks of 5 instead of 50. This is a
temporary fix to the console not working issue.
-Took out super matter bin's reliability base initialization call. It's
a goddamn container, how does it break?
*bumped super cap cells reliability base up a bit. these should now take
less time to produce working versions.
+Added a comment to the stacking machine file to denote where I think
the console buggers itself. somebody try and dick around with this.
+attempted to make it so the owl mask will change your voice to "The
Owl"
+added an HoP closet that spawns all that owl stuff
-removed the rapid syringe gun and the emp grenades from HoP quarters.

BUGS:
The ruskie shuttle does not move the way it's supposed to.
The computer that controls said shuttle appears to be accessible by
anyone, when it's supposed to be locked with centcom_general access.