Rip through Admin
I was going through the admin powers trying to understand how they are issued. Moved some
around a bit but they all work the same.
I want to push this to master because everything seemed to break again on me. I want to see
if others get this problem. If the map borkes just rollback the commit, I won't mind