mooooooooooar
Did a lot more to rockstation. It's missing a security area and a medbay
and lots of required machinery/transit stuff (like wires and pipes).
Once those are in, we should be good to go. >:U

Also there is a set aside mime and clown area even though we don't have
those. (yet...)