Whatever blarg
Not sure what I can say I did on this. I did fix slimes, the edit I made before made it
impossible to put them on tables for cutting open. Tested and fixed and still uses optimized TG
code. Also working on teststation map which I'm making small and to be a little bit of
everything.