Small commit for fire
After seeing that there was something wrong with the fire code from TG, I asked
Max about it and he told me what I was missing. Added the needed extras for
things that can make you flameable. This included a small update to the types of
bullets you can shoot.