Telescience added
Added files for Telescience. Map needs a place for it eventually but I can't do it. Also added
and updated some other files that were needed to make telescience work. Messed a little more
with slimes and they are still not done. Also removed what I think is all the .bak files.