Rockstation!
It is out! Rockstation is now playable and is set as the current play
map in the build. Please be aware that there are some fine tunings
needed. Note the following:

-THERE IS NO EFTPOS SYSTEM IN PLACE. The disposal system will route
everything to a place in QM, which will then go to the disposal room to
be ejected.
-Make note of any doors to nowhere* you find and tell me if you find
one.
-Make note of anywhere that is missing a fire alarm or air alarm
-There are a few areas missing from Rockstation that are there in the
TGstation default map. Raise hell if you think they need to be in.
-Note anything else that is obviously a problem, such as inability to
access the right places or areas accidentally exposed to space.

*A door to nowhere is my term for when I place a door and forget to put
the turf under it to a floor tile. This can be a wall at best or a space
tile at worst. Either will screw with gameplay moderately.