Challenge removal and other goatfuckery
The main and biggest point of this push was the removal of challenge
from the randomZ list for now. It was causing the game to tick at
painfully low speeds. For example, Atmos and the singularity moved quite
visibly slower. Arcticwaste and Academy are still in, however.

Now as follows is the wall of other shit I did:

*moved the QM's filing cabinet so it was within reach of him/her.
+added a bunch of newscasters
*bumped supply shuttle move time up a minute
+added some light switches
+added a maintenance door to primary tool storage
+added a few text easter eggs
+added a Pingas Wheel generator. This does the exact same thing as the
Magic Pacmans from old terra. Also added two spawns for them.
+added air alarms to mining dock and bar.
+added a few scrubbers to the hall
+added a tech storage room. Holy derp, how did I miss this?!
-Removed a chair that was placed stupidly in arrivals.
*changed some tinted windows into normal ones, now that I know what they
do differently.
*made the otherwise-difficult-to-find science wing a bit more obvious
+fixed the lack of lights in the xeno/virology halls
+added another solar array. This one behaves differently and is wired
with blue wire instead of red.
+added another external access airlock contraption
+added a conveyor to QM to cover for something I thought would work but
doesn't
+added a microwave to kitchen. >:|