Bluh bluh
*fixed Shiper, which got buggered in one of my merges. It functioned
properly (except for a syntax change that I fixed), but the icons were
horribly broke. Now it's all good.
~+ did a bit more work on Rock2. IF ANYONE FEELS LIKE DRAWING UP A NEW
ATMOS FEEL FREE* TO DO SO BEFORE I MAKE ANOTHER COMMIT, SINCE I DON'T
WANT TO TOUCH IT.
^^^ (*yeah the fuck right who's gonna do that)

But on that note if anyone feels like it ram an atmos section to the
right of the engine hall. I have pipes leading to it already and I just
want them to connect there. Don't worry about having anything open to
space, I'll fuck with that myself.
If nobody does it by the time I get to it I have a dumbass plan that
might work ok