niggapush
push for cippie because blargage
added musics but didn't fuck with lobby list yet.
Also added some terrastation areas in a separate file so that the
merging doesn't fusk with it and saves us headaches down the road. Some
are old and unreimplemented but some are new for the new map