Map Tweaks
Reworked assembly line to be less abandoned.  Created Assembly access ID
and gave it to: Mechanic, Station Engineer, Chief Engineer, Roboticist,
and Research Director.

Wired all the solar arrays and upgraded northeast solar array. Buffed
the SMES's in engineering to start with and store more power. This is to
make a engineering an easier task for our usually small server
population.

Changed beaker volume for small/large to please Doc.