HAVE FUN WITH THIS
Got a semi-complete map of Ministation to work. All turfs should be there as well as areas. However all items have been set to their default setting (ie: all cables and pipes lost their directions and doors their special names and id restrictions). All supply and scrubber pipes and power cables have been removed due to this. The rewiring and piping is not hard as I got it done previously in a day. I'm not doing it again.
Also to note, some objects did not save over and are missing. Consult a clean version of Ministation.dmm from /tg/ code to fix these issues. Mobs will also need a review but only Ian, Pun Pun, and a few mice were on the station I believe. In the meantime I will work on getting a version of z2.dmm that will work.