Progress push #1
Rock2's Zlevels partially fixed; this one fixes 2, 3, and 6. (2 & 3 were
just copypaste as they are only in so the map will function and are no
different from cyberiad. Z6 was the actual workload here, but so far the
easiest of the ones I need to fix.)

Pushing only so I have a fallback if something horribly buggers further
down.