OG update
Okay, 90% of cyberiad has been copy/pasted in for reference, so I don't
have to have both files open in DM at the same time. The only thing that
isn't copied are the maintenance tunnels, as OG has a very different
layout and each room will eventually be reoriented and condensed.

Probably going to mass-delete all of the piping and wiring because it's
going to be ass otherwise.