keycards
Moved the HoP, NT rep, and medbay Keycard Authentication Devices to
places that are more intuitive to use properly.
The same device on the bridge has had it's logical location moved to
make it easier to use, but its physical location is pretty much the
same.

Also, WHERE THE FUCK DID THAT MERGE COMMIT WITH USELESS CONFIG COME
FROM?! I did NOT make that commit!