Construction
The station is undergoing some structural changes, pardon the dust!
Three new pod bays have been added to various places around the station.
In addition, the mechanic's workshop has been expanded to cover the
potential increase in demand.
On a totally unrelated note, the pool has been temporarily drained.
Also, if anyone finds the clown that replaced the boxing ring's floor
wil bananium, please alert the local authorities. Construction workers
have moved the large table in the rec room. And by moved, we may or may
not mean that the wood was cannibalized for use in other parts of the
project. We will have a replacement ready as soon as possible. All the
items that were on the table before are still there, just moved.
Keep up the good work!