Errors Corrected
Following reports from what is totally a vast number of people and not
the whining of a single individual that is the only voice we care about,
it has come to our attention that one of the installed podlocks is
completely inaccessible to station employees intending to utilize the
relevant equipment it is designed to house. We have strongarmed some ERT
members to take a block ram to the surrounding area to widen the passage
into this area during their off hours. The pod bay in question should
now be fully accessible.