Errors corrected
Fixes map errors in Rock2. Also "fixes" code errors.

Known problems:
-finds_misc.dm, finds_special.dm, and reconstitutor.dm have
commented-out lines to fix compile errors. This is especially unnerving
in finds_special, as there is a line that removes between 25 and 50
blood from the target by way of editing a "vessel" variable on the
target that the method gets called against. This variable doesn't
presently exist and presumably used to and I am not sure at the moment
what this has changed to.
-Some Xenoarch files have been unticked entirely to avoid compile
errors. For the most part, something changed in the code somewhere else
and these files need to be reworded and I don't presently know what the
new code is.
-Rock2, z2, z4, and z5 had errors before this push. All but z2 were
fixed by way of mass text edit. For whatever reason, attempting to do
the same to z2 resulted in errors involving mysterious ASCII characters
appearing where they shouldn't. Repathing natively in Dream Maker,
miraculously enough, fixed this. So it works now, but I felt like
dropping this here to make note.