fuck
So of course that last commit broke shit, which was part my fault for
not catching it. I kinda-sorta reverted the changes here, but the code
is still viewable as comments next to where it's supposed to go. I also
kept the new files.
I/someone need/s to find the commit that introduced the vars this commit
evidently relies on and either nab that commit also or port it so it
works with what's here

fuck my life ugh