bugfixes and modifications hooray
*moves the public chem fridge to somewhere that's not in the middle of
the goddamn freeway
*replaced said fridge with the I think correct object (this should not
have any material effect on the game outside of fluff)
+added a door switch to exit medbay
*moved the existing buttons behind the medbay reception desk and
hopefully fixed the fact they don't always work
*suplexed the robotics desk door
*moved some stuff around on z2. The sol station and solarium should now
both be able to see the sun from their respective windows, but unable to
see each other. I moved some other things on that z-level around to
accomodate this movement. This shouldn't affect anything else.
*Edited the following salvage templates to have more stuff:
{s2, s3, s4, s6, s8}
+added a new salvage template to the smol list.
+added an area to better handle cramming people up my ass (they can now
be send-ed there and always pop up in the same spot with admin fuckery.
This solution is janky and lame I know but until I get a proper verb for
this this is an easy pseudofix)
*moved trade shuttle so it hopefully locks that damn airlock at game
start