Secondary tactical anal checkpoint
+added a get out frog. When clicked with an empty hand, it will make you
get out (of my colon)
+added icons for said froge and the RBD, a device not yet implemented
that will be a possible prize at the end of the arsehole cave.

With this update, it is now possible to leave my arse if rammed up it.
There is still no real benefit to doing this, but at least you won't be
stuck there if you do go.