Quality of Life changes
+fixes a missing pipe in engine room
*fixes the salvage captain's fax machine, now he or she can actually
fukken access it.
*tweaeaeaked the cooldown system for emotes and made it so it will allow
emotes called in certain situations to ignore the cooldown. Currently,
this is only used by the Fart spell and Fartonium.
*Rebound ctrl+f in hotkey-disabled mode and F in hotkey mode to
Say"*fart. Now there's a fart button, lads. Rejoice. (Thanks a bazillion
to Qui for figuring this out!)
+added another Fax machine to the NT Rep's office. Now they have two:
one for the local network, one for centcom. I think that's how this
works.
+added a line of text to the fart spell. Now it identifies the caster,
in a humorous way.