Minor bug fixes
This pushes some finishing touches before the merge.
Note that research borg has been rendered unusable for the time being by removing it from the selection of modules. There will be a part two of this PR at some point to finish what was started here. The new code for the borg module is still in and compiles, but since I only needed to remove that one word to make it inacessible I felt no need to remove or comment it out.

Co-Authored-By: bastgg <bastgg@users.noreply.github.com>